+++ CHANGELOG.md
@@ -1,10 +1,15 @@
 # @gerardpastor/lexer
 
-## 0.1.4
+
+## 0.2.0
+
+### Minor Changes
+
+- d1a5b84: Refectored lexer
 
 ### Patch Changes
 
-- 6669084: Fixed use of value group
+- 235a84d: Added next valid and word boundary
 
 ## 0.1.3